refactor: 팀원 피드백 반영
1. jwt secretKey application-secret으로 이롣
2. 통계 API Response Message 변경
3. 통계 API Controller Parameter required true 추가
4. 통계 API 예외 발생 시 CommonResponse 응답으로 변경
5. 통계 API QueryDSL에서 count의 리턴타입과 x_count의 리턴타입 int로 통일했던 부분->long으로 통일하도록 변경,DTO에서 받아주는 타입 long으로 변경

+++ src/main/java/com/wanted/sns_feed_service/feed/repository/FeedRepositoryImpl.java
@@ -134,7 +134,7 @@ public class FeedRepositoryImpl extends Querydsl4RepositorySupport implements Fe
         //날짜/시간 형식 설정
         StringTemplate formattedDate = getFormattedDate(feed.createdAt, type);
         //count 일 시 게시물 개수 일자별 출력, 아닐 때는 view count 등 합계 출력
-        return select(Projections.fields(StatisticsResponse.class, formattedDate.as("date"), value.equals("count") ? feed.count().intValue().as("count") : value.equals("view_count") ? feed.viewCount.sum().as("count") : value.equals("like_count") ? feed.likeCount.sum().as("count") : feed.shareCount.sum().as("count")))
+        return select(Projections.fields(StatisticsResponse.class, formattedDate.as("date"), value.equals("count") ? feed.count().as("count") : value.equals("view_count") ? feed.viewCount.sum().longValue().as("count") : value.equals("like_count") ? feed.likeCount.sum().longValue().as("count") : feed.shareCount.sum().longValue().as("count")))
                 .from(feed)
                 .where(hashtagFilter(hashtag),
                         feed.createdAt.between(start.atStartOfDay(), end.plusDays(1).atStartOfDay().minusNanos(1))
